Add only trusted projects' bin directory to $PATH
Assuming the binstubs for a project are in the local bin/ directory, you
can even go a step further to add the directory to shell $PATH so that
rspec can be invoked without the bin/ prefix:

    export PATH="./bin:$PATH"

However, doing so on a system that other people have write access to
(such as a shared host) is a security risk.

https://github.com/sstephenson/rbenv/issues/309

Put this in `zprofile` because Arch (and Fedora/Debian) puts `PATH`
manipulation in `/etc/zsh/zprofile`, which is sourced after `~/.zshenv`
(our preference). We chose `~/.zprofile` because we were sick of getting
command-not-found errors in non-interactive settings.

If the system-level zprofile (`/etc/zsh/zprofile`) sets a `PATH`, it
will clobber anything you do in `~/.zshenv` since it's sourced after. We
also like that `~/.zprofile` is sourced less frequently so moving what
you can there will speed up terminal launch, by however small an amount.

It should not be `zlogin` because:

http://zsh.sourceforge.net/Intro/intro_3.html

> `.zlogin` is not the place for alias definitions, options, environment
> variable settings, etc.; as a general rule, it should not change the
> shell environment at all. Rather, it should be used to set the
> terminal type and run a series of external commands (fortune, msgs,
> etc).